⚙️ Refactor navbar items, separate options
